refactor: Extract UsernameInput and AvatarPicker as component
