+++ src/translations/en.json
@@ -193,7 +193,10 @@
       "theme": "Theme"
     },
     "title": "General",
-    "usernameLength": "Username must be at least 3 characters."
+    "username": {
+      "error": "There was an error updating your username.",
+      "length": "Username must be at least 3 characters."
+    }
   },
   "requireAuth": {
     "cancel": "Cancel",
